Reverted re-clustering logic
This commit:

  commit 472fe2ebc36a73c02f3960d5263b81e1f0340638
  Author: Dima Kogan <Dmitriy.Kogan@jpl.nasa.gov>
  Date:   Wed Nov 1 20:53:13 2023 -0700

    After finding each plane I re-cluster before finding new planes

    Otherwise I could possibly find planes that span multiple clusters

The general idea is sound: we need to re-cluster the point cloud after a chunk
has been removed from it. But this implementation was overly-complex and
bug-prone. And there was at least one bug: the same cluster/subcluster IDs were
reported to be processed multiple times. At the very LEAST there was a missing
i_subcluster+=1. But I simply remove that logic, to be re-added later in a
better way.