non-core imports AFTER parse_args
This makes --help and bad-option-parsing be fast, avoiding imports that it
doesn't need. The common block:

  import sys
  import argparse
  import re
  import os

is my convention (see all mrcal tools, for instance). We don't always need ALL
of these, but it's so common, I find it a good idea to always have this set
available. I always want to be able to sys.exit() and have it work. In any case,
this structure is how all my tools are set up, and I'd like to keep it that way,
if it's ok.