Reordered imports to make testing easier
I want to import the local mrcal. The install mrcal is fine in the Python layer,
but I want the new functions in the C layer. Importing clc before mrcal makes
sure to load the local libmrcal.so (due to the RPATH), so I don't need to worry
about the PYTHONPATH