Rearchitected stage2 (segment clustering)
Previously I was looking for a connected component. This allowed me to find
complex shapes; and stage3 then had to handle complex shapes, which was unideal.

We never are actually looking for complex shapes: we want a big, convex, flat
thing, so stage2 now:

- Searches laterally in both directions to find the segments on the plane in
  this ring
- Finds one matching segment in the next ring (one direction)
- Loop from that one matching segment

This is simpler, and produces simpler output for stage3 to process. The stage3
code is NOT yet updated to do anything differently; it currently does the same
thing with the new data structure. Once this is debugged and working, stage3
will be updated as well